code-art72/tub/requirements.txt
Django==1.3
gondor==1.0b1.post14
psycopg2==2.4.1
PIL==1.1.7
South==0.7.3
sorl-thumbnail==11.09.1
python-memcached==1.48
django-registration==0.7
django-subdomains==1.1.2
